Move UploadData to src/workspace-data
